Cards/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Hero, Monster, Skill


class SkillSerializer(serializers.ModelSerializer):
    class Meta:
        model = Skill
        fields = '__all__'


class HeroSerializer(serializers.ModelSerializer):
    skills = SkillSerializer(many=True, read_only=True)

    class Meta:
        model = Hero
        fields = '__all__'


class MonsterSerializer(serializers.ModelSerializer):
    skills = SkillSerializer(many=True, read_only=True)

    class Meta:
        model = Monster
        fields = '__all__'