pyproject.toml
[project]
name = "cardgame"
version = "0.1.0"
description = ""
authors = [
    {name = "strahser",email = "ssg-81@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
